test-requirements.txt
pyhamcrest
pytest
unidecode==1.0.23
